Revise readme, fix bugs.

+++ mmle/distributed/distributed_c10d.py
@@ -14,6 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+import functools
 from tempfile import gettempdir, NamedTemporaryFile
 
 import torch  # pylint: disable=E0401
@@ -25,20 +26,25 @@ from torch.nn.parallel import DistributedDataParallel as DDP  # pylint: disable=
 import mmle.utils as mut
 
 
+def _spawn_wrapper(world_size, args, kwargs, rank, fn, file_name):
+    distd.init_process_group(
+        "nccl", init_method=f"file://{file_name}", rank=rank, world_size=world_size
+    )
+
+    if args is None:
+        args = ()
+    if kwargs is None:
+        kwargs = {}
+    fn(*args, **kwargs)
+
+
 def spawn(fn, args=None, kwargs=None, nprocs=1, join=True, daemon=False):
-    def _spawn_wrapper(rank, fn, file_name):
-        distd.init_process_group(
-            "nccl", init_method=f"file://{file_name}", rank=rank, world_size=nprocs
-        )
-
-        if args is None:
-            args = ()
-        if kwargs is None:
-            kwargs = {}
-        fn(*args, **kwargs)
-
     mp.spawn(
-        _spawn_wrapper, (fn, NamedTemporaryFile().name), nprocs=nprocs, join=join, daemon=daemon
+        functools.partial(_spawn_wrapper, nprocs, args, kwargs),
+        (fn, NamedTemporaryFile().name),
+        nprocs=nprocs,
+        join=join,
+        daemon=daemon,
     )
 
 
@@ -54,7 +60,8 @@ def sync(module, rank=None):
 
     module.load_state_dict(torch.load(tmp_p, map_location={"cuda:0": f"cuda:{rank}"}))
 
-    tmp_p.unlink()
+    if rank == 0:
+        tmp_p.unlink()
 
 
 def distribute_module(module, rank=None, *args, **kwargs):
